requirements.txt
asgiref==3.8.1
boto3==1.39.8
botocore==1.39.8
certifi==2025.7.9
charset-normalizer==3.4.2
Django==4.2.13
django-environ==0.12.0
django-storages==1.14.6
djangorestframework==3.15.1
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
jmespath==1.0.1
packaging==24.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.0
typing_extensions==4.12.1
uritemplate==4.1.1
urllib3==1.26.20
